Update patch set 2
Patch Set 2:

> Patch Set 2: Published edit on patch set 1.

Hmm... I was still thinking about this myself, and I had a different direction in mind, that I was hoping to try tonight or tomorrow night. My aim would be to keep the conservative governor when the screen is off, but keep interactive when screen is on. Would you mind waiting for a first patch?

Patch-set: 2